Maps Release 1.4.5 - 13 December 2024
Please note that the Montlake Lid bike/ped bridge and interconnection area are NOT YET ACTUALLY OPEN as of today, December 13th. It was supposed to be. Opening ceremony is December 14th (tomorrow) and it is finished enough _to_ open tomorrow - and hopefully will - but landscaping is incomplete and I suspect there may be temporary closures for landscaping purposes.